Change Array[] to List<>
Each size declared in the // VARIABLES // section of 'Gui.java' has changed its class to List<Dimension>. Also, there is a getter for each size.

Applying this to the other classes, sizes are set in the constructor by calling the getters from 'Gui.java', or in some cases customazing them with 'Arrays.AsList()'. This also applies to 'gridBagConstraintsCentered'.

Furthermore, the methods for setting the size of each type of component have been unified in a public 'setComponentSize()' method that uses 'if (obj instance of X) {} else if ([...]) {}' to choose from a set of private 'setXSize()' methods.

These changes have only been applied to classes 'Gui.java', 'MenuPanel.java', 'SettingsPanel.java' and 'ScreenPanel.java'.
Both 'ScreenPanel.java' and 'DisplayPanel.java' need a deep remodelation...