Added one slider to 'GameOfLifeOptionsPanel.java'
This slider wasn't just one JSlider component, but the addition of 2 JLabels (name of the attribute and real-time value). In the process, it was found that most part of the methods could be written in the abstract parent class.

All of this has only translated into one 'generateSlider()' method, but the workflow can make new additions of this kind of methods way simpler.

Oh, I almost forgot. Class 'ScreenPanel.java' has been renamed to 'GameOfLifeOptionsPanel.java'