Attribute 'maxLife' changed for 'cellColor'
For class 'GameOfLifeDisplayPanel.java', instead of putting a 'maxLife' attribute to give the cells a range of life states, it has been put a 'cellColor' attribute to set the color of cells.

As small as this change is, the commit is done mainly to save from just before applying the next change.